+++ lib/src/search_app_bar_page/controller/searcher_base_control.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:get/state_manager.dart';
-import 'package:search_app_bar_page/src/search_app_bar_page/controller/utils/filters/functions_filters.dart';
 
 abstract class SearcherBase<T> {
   VoidCallback? onCancelSearch;
@@ -33,10 +32,6 @@ abstract class SearcherBase<T> {
   void sortCompareList(List<T> list);
 
   bool? sortCompare = true;
-
-  bool isAsync = false;
-
-  ListAsync<T>? listAsync;
 
   //FutureOr<List<T>> Function(Object error)? onAsyncError;
 }