+++ lib/src/search_app_bar_page/ui/core/search_app_bar/search_app_bar.dart
@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:search_app_bar_page/src/search_app_bar_page/controller/searcher_base_control.dart';
+import 'package:search_app_bar_page/src/search_app_bar_page/controller/searcher_page_controller_variable.dart';
 import 'package:search_app_bar_page/src/search_app_bar_page/controller/utils/filters/functions_filters.dart';
 import 'package:search_app_bar_page/src/search_app_bar_page/ui/core/search_app_bar/search_paint.dart';
 import 'package:search_app_bar_page/src/search_app_bar_page/ui/core/search_app_bar/search_widget.dart';
@@ -157,6 +158,9 @@ class _SearchAppBarState<T> extends State<SearchAppBar<T>>
 
   void cancelSearch() {
     widget.controller.isModSearch = false;
+    if (widget.controller is SearcherPageControllerVariable<T>) {
+      (widget.controller as SearcherPageControllerVariable<T>).onSearchList([]);
+    }
     widget.controller.rxSearch.value = '';
     _elevation = widget.elevation;
     // Reset both animations to initial state (AppBar fully visible, no ripple)